Updating some samples to use Fibonacci
